APIControllerDelete - update deleteUser()

+++ src/Controller/APIControllerDelete.php
@@ -21,16 +21,26 @@
 			}
 		}
 		
+		public function deleteUser($id, Application $app) {
+			if ($app['dao.user']->get($id)) {
+				$app['dao.depense']->deleteByUser($id);
+				$app['dao.user']->delete($id);
+				return $app->json(array(
+					'status' => 'OK'
+				), 200);
+			}
+			else {
+				return $app->json(array(
+					'status' => 'KO',
+					'error' => 'Pas d’utilisateur enregistré avec l’id '.$id
+				), 400);
+			}
+		}
+		
 		public function deleteGroup($id, Application $app) {
 			$app['dao.depense']->deleteByGroup($id);
 			$app['dao.user']->removeFromGroup($id);
 			$app['dao.group']->delete($id);
-			return $app->json('No Content', 204);
-		}
-		
-		public function deleteUser($id, Application $app) {
-			$app['dao.depense']->deleteByUser($id);
-			$app['dao.user']->delete($id);
 			return $app->json('No Content', 204);
 		}
